[pylint.message] Create a MessageDefinition.check_msgid function
